Major change in handling baseline surveys
The old `hh.baseline.sample` column was incorrectly identifying households that received a baseline survey. In fact, at this point we are only able to satisfactorily link baseline survey respondents to census households/individuals in wave 2 (Kakamega), using `person_key` in the sample list (in the data folder) and `person` in the baseline survey data. The `person` IDs for wave 1 have no corresponding IDs in the sample list. It looks like the sample list file was updated for wave 2 and overwrote the IDs that were used to link the sample list to the baseline data. I tried to do some work linking rows by `cluster.id`, `village` and 'hhh_last_name` but I'm still left with ~200 unidentified rows and a small number of duplicates. The way this whole survey was done was a mess. In fact, for wave 2, I still have to use `person_id` to link to the proper row in the census data (I still end up with 23 wave 2 respondents that I can't link to the census data).